+++ apps/builder/app/builder/features/style-panel/property-label.tsx
@@ -29,7 +29,7 @@ import type {
   StyleValueSourceColor,
 } from "~/shared/style-object-model";
 import { useComputedStyles } from "./shared/model";
-import { StyleSourceBadge } from "~/builder/shared/style-source";
+import { StyleSourceBadge } from "./style-source";
 import { createBatchUpdate } from "./shared/use-style-data";
 import { $virtualInstances } from "~/shared/awareness";
