+++ apps/builder/app/builder/features/style-panel/style-source-section.tsx
@@ -25,7 +25,7 @@ import {
   type ItemSource,
   StyleSourceInput,
   type StyleSourceError,
-} from "~/builder/shared/style-source";
+} from "./style-source";
 import {
   $registeredComponentMetas,
   $selectedInstanceStatesByStyleSourceId,
